fix(git): ensure chanined execution stops when `__gbp` is terminated abruptly
On some OSes (e.g. Ubuntu), abruptly terminating the process via `Ctrl+C` (or
other similar method), exits the current sub-process (with a 0) and not the
parent process. If this happens while `__gbp` is being executed, the following
commands will continue to execute, which is undesirable.

This commit fixes it, by detecting a exit (with 0) while the `__gbp` prompt is
still visible and manually exiting the process with 1.