In 5140c35eb, `gcoghpr` was updated to allow pushing to a remote even when the `gcoghpr` remote alias was overwritten (e.g. by a subsequently `gcoghpr`'d PR). This was achieved by not creating a remote alias and setting up branch tracking manually (using the origin URL).
However, it turns out that with this setup it is not possible to update
an existing commit and force-push (getting "stale info" error).

This commit fixes the issue (while still preserving the ability to push
changes in previously checked out PRs) by bringing back remote aliases,
but using the PR author name in the remote alias (thus ensuring there
will be a unique alias per author).

NOTE:
This new approach creates remote aliases that need to be manually
cleaned up (when no longer needed), but this is worth the extra overhead
in order to be able to update existing commits. Also, by using a common
prefix for all remote aliases, it will be easy for the user to spot
these auto-generated remote aliases (and remove them).