feat(node): use `http-server` instead of `light-server` in `srv` aliases
This commit changes the `srv` alias to use `http-server` instead of
`light-server` under the hood. This means that the available options are
now different.

Reasoning:
`http-server` generally works better out-of-the-box and is more powerful
than `light-server`. `light-server`'s only benefit is its built-in
watching capabilities, but these arenot relevant for the `srv` alias.

NOTE:
The `srvw` alias has not been changed and still uses `light-server`.