fix(git): do not unnecessarily check out master in `gcoghpr` alias
Previously, the `gcoghpr` alias would try to check out the master
branch. This is not necessary for the rest of the operations to work and
could cause failures when the master branch was already checked out in
another [git worktree](https://git-scm.com/docs/git-worktree).

This commit removes the step that checks out master from the `gcoghpr`
operations.