fix(git): correctly check out PR when already on the target branch
In the past, `gcoghpr` would always check out master before creating and
switching to the local branch for the specified PR. In commit
5637349ddc319b239f4b8f6993661dbc6b57526e, this step was removed as
unnecessary.

It turns out that checking out master (or another branch) is necessary
if we are already on the target branch. This commit conditionally adds
back the step, only when needed (i.e. only when already on the target
branch).