fix(git): correctly display colors in `__g-pick-branch`
In a recent commit, `__g-pick-branch` was updated to use `chalk` for
styling the branch names shown in `__g-pick-branch`. It turns out that
this breaks coloring of the output in `inquirer.prompt()`.

This commit fixes this by removing the use of `chalk` for branch names.
To still allow distinguishing `gcoghpr-` branches, this commit updates
the formatting of such branches' names.