fix(node): fix `nup`, `nud`, `nua`, `yup`, `yud` and `yua` aliases on *nix
This commit fixes the `nup`, `nud`, `nua`, `yup`, `yud` and `yua`
aliases on *nix platforms by ensuring no backticks (`` ` ``) are used in
the Node.js command that is evaluated directly in a shell.

It also adds a note to similarly affected functions to warn against the
use of backticks in the future.