fix(git): detect and use default branch in aliases
Previously, it was assumed that `master` is the default branch and that
was hard-coded in aliases that required to use it.

This commit updates all affected aliases to correctly detect and use the
actual default branch.

NOTE:
In lieu of a reliable and efficient way (i.e. without requiring network
access) of detecting the default branch, we are currently looking for
the existence of some common default branch names, namely `master` and
`main`.