fix(git): properly detect default git branch in `gdefb` alias (if configured)
If `origin/HEAD` is set on the git repository, allow `gdefb` to use to
properly detect the actual default branch, instead of relying on the
heuristic of checking for the existence of the `main` or `master`
branches.

If `origin/HEAD` does not yield anything, fall back to the old
heuristic, but prioritize `main` over `master`.

If `origin/HEAD` is not set in the local working copy (but the `origin`
remote does exist), the user can set it with the command:
`git remote set-head origin --auto`