fix(utils): make `importWithEnv()` more robust and versatile
The previous implementation of `importWithEnv()` did not work on some
versions of Node.js for certain packages (e.g. inquirer@9.3+) that were
not compatible with `require()` (and thus neither with
`require.resolve()`).

Make the `importWithEnv()` function more robust and versatile, by
allowing the caller to pass in the import function and only dealing with
calling it "in" the desired environment. This allows the caller to
decide what import method to use (e.g. `import()` vs `require()`),
if/how to resolve the import path, etc.